pipeline/kafka_consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
redis==5.0.4
